Add files via upload
Here I've coded an ab initio Python script for constructing a sto-3g basis set and solving the Roothaan-Hall equations for a general molecule, using water as an example. For computing the 4-index electron repulsion integrals (used to form J and K), I sent the task to an external server (RunPod H100 PCLe) so that this 4D matrix is stored just before Section 6 in the script. The basic script for finding J and K manually is still here, it just takes about 5 hours to run on a home computer.